apps/python-api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
httpx==0.28.1
beautifulsoup4==4.13.4
dnspython==2.7.0
anthropic==0.52.0
pydantic-settings==2.9.1
python-dotenv==1.1.0
asyncpg==0.30.0
playwright>=1.49.0
